[Runner] Update and improve the runner
Most impactful change:
Now we inter intersperse calls to `ray memory` in between cells.
This should be better than querying it in parallel because we weren't
sure if we'll "catch" the peak usage because of the latency of
`ray memory`. That is, until `ray memory` returns a result, the
execution might have moved to a different cell, or finished altogether,
and we might miss it. Because now it's interspersed, it can take
as long as it wants and we'll wait before going to the next cell.

As I suspect that `ray memory` will add overhead, we do different runs
when we call it, to measure memory, and different runs to measure
time.

Other changes:
- Use rewriter as a library. We don't have to go through magic functions.