[Runner] Remove to_py.py
The original idea of this file was to translate the .ipynb notebooks
to pure .py and run that, instead of going the many scripts
that may add overhead, and also executing with IPython.

However, we anyway need to use ip.run_cell() in the .py too and so
it seems to me that it would basically be the same solution.

Note: There is an alternative to get pure Python. Run the pattern
matcher "offline" (and measure rewriting overhead there) to get
the rewritten cells. Then, paste those into a file that is pure
Python.

This would take care of most of the problem, but still apply()
needs to rewrite code internally that it then runs with ip.run_cell().
We could run the notebooks once to get this rewritten version
of apply() too, paste those into the .py and eventually run that.

However, I think this would create quite an unrealistic setting.